# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/superadmin/register": {
      "post": {
        "description": "Crea una nueva cuenta de superadministrador en el sistema",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterSuperadminDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Superadministrador registrado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "user": {
                      "id": "uuid",
                      "email": "admin@vekino.com",
                      "name": "Juan Pérez",
                      "role": "SUPERADMIN"
                    },
                    "session": {
                      "token": "session_token",
                      "expiresAt": "2024-12-31T23:59:59.000Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos de registro inválidos"
          }
        },
        "summary": "Registrar un nuevo superadministrador",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/superadmin/login": {
      "post": {
        "description": "Autentica un superadministrador y establece una sesión",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginSuperadminDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inicio de sesión exitoso",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "user": {
                      "id": "uuid",
                      "email": "admin@vekino.com",
                      "name": "Juan Pérez",
                      "role": "SUPERADMIN"
                    },
                    "session": {
                      "token": "session_token",
                      "expiresAt": "2024-12-31T23:59:59.000Z"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credenciales inválidas"
          }
        },
        "summary": "Iniciar sesión como superadministrador",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Vekino",
    "description": "API para gestión de condominios",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de desarrollo"
    },
    {
      "url": "https://api.vekino.com",
      "description": "Servidor de producción"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Token de autenticación JWT",
        "in": "header"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Cookie de sesión de Better Auth"
      }
    },
    "schemas": {
      "RegisterSuperadminDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Correo electrónico del superadministrador",
            "example": "admin@vekino.com"
          },
          "password": {
            "type": "string",
            "description": "Contraseña del superadministrador (mínimo 8 caracteres)",
            "example": "password123",
            "minLength": 8
          },
          "name": {
            "type": "string",
            "description": "Nombre completo del superadministrador",
            "example": "Juan Pérez"
          }
        },
        "required": [
          "email",
          "password",
          "name"
        ]
      },
      "LoginSuperadminDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Correo electrónico del superadministrador",
            "example": "admin@vekino.com"
          },
          "password": {
            "type": "string",
            "description": "Contraseña del superadministrador",
            "example": "password123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
    }
  }
}